requirements.txt
setuptools>=70.0.0
alembic>=1.15.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.10
asyncpg>=0.27.0
alembic>=1.11.0
sanic>=24.6.0
Jinja2>=3.1.2
pydantic>=2.0
PyJWT>=2.7.0
